Next gen test program; template files for monitoring components

+++ bin/test-fb-image.sh
@@ -110,33 +110,73 @@ function doitall {
 
    #arg1 Full container image
    #arg2 name of template file
+   #arg3 prefix to insert in placeholders
+   #prefix="_"
    parseFullImage "$1"
+   prefix=${3:-""}
 
    tempfile="/tmp/container_image.yaml"
+   template_file=$2
 
-   rm -f  $tempfile
-   cp $2  $tempfile
+   if [ "$template_file" != "TEMPFILE" ]; then
+      rm -f  $tempfile
+      cp $template_file  $tempfile
+   else
+      echo "DEBUG: modifying existing file"
+   fi
 
    if [ "$AIRGAP_DEPLOYMENT" == "true" ]; then
       REGISTRY="$AIRGAP_REGISTRY"
    fi
-
-   v4m_replace "__GLOBAL_REGISTRY__"    "$REGISTRY"                 "$tempfile"
-   v4m_replace "__IMAGE_REPO__"         "$REGISTRY\/$REPOS\/$IMAGE" "$tempfile"
-   v4m_replace "__IMAGE_TAG__"          "$VERSION"                  "$tempfile"
-   v4m_replace "__IMAGE_PULL_POLICY__"  "Always"                    "$tempfile"
-   v4m_replace "__IMAGE_PULL_SECRET__"  "null"                      "$tempfile"       #Handle Single Image Pull Secret
-   v4m_replace "__IMAGE_PULL_SECRETS__" "[]"                        "$tempfile"       #Handle Multiple Image Pull Secrets
+   v4m_replace "__${prefix}IMAGE_REGISTRY__"     "$REGISTRY"                 "$tempfile"
+   v4m_replace "__${prefix}GLOBAL_REGISTRY__"    "$REGISTRY"                 "$tempfile"
+   v4m_replace "__${prefix}IMAGE_REPO__"         "$REGISTRY\/$REPOS\/$IMAGE" "$tempfile"
+   v4m_replace "__${prefix}IMAGE__"              "$IMAGE"                    "$tempfile"
+   v4m_replace "__${prefix}IMAGE_TAG__"          "$VERSION"                  "$tempfile"
+   v4m_replace "__${prefix}IMAGE_PULL_POLICY__"  "Always"                    "$tempfile"
+   v4m_replace "__${prefix}IMAGE_PULL_SECRET__"  "null"                      "$tempfile"       #Handle Single Image Pull Secret
+   v4m_replace "__${prefix}IMAGE_PULL_SECRETS__" "[]"                        "$tempfile"       #Handle Multiple Image Pull Secrets
 
    cat $tempfile
 
 }
 
 echo "*****************"
+
 doitall "$FB_FULL_IMAGE"          "logging/fb/container_image.template"
+
 doitall "$OS_FULL_IMAGE"          "logging/opensearch/os_container_image.template"
+doitall "$OS_SYSCTL_FULL_IMAGE"   "TEMPFILE"  "OS_SYSCTL_"
+
 doitall "$OSD_FULL_IMAGE"         "logging/opensearch/osd_container_image.template"
+
 doitall "$ES_EXPORTER_FULL_IMAGE" "logging/esexporter/es-exporter_container_image.template"
+
+GRAFANA_FULL_IMAGE="docker.io/grafana/grafana:10.2.1"
+GRAFANA_SIDECAR_FULL_IMAGE="quay.io/kiwigrid/k8s-sidecar:1.25.2"
+doitall "$GRAFANA_FULL_IMAGE"     "monitoring/grafana_container_image.template"
+doitall "$GRAFANA_SIDECAR_FULL_IMAGE"   "TEMPFILE"  "SIDECAR_"
+
+ALERTMANAGER_FULL_IMAGE="quay.io/prometheus/alertmanager:v0.26.0"
+GRAFANA_FULL_IMAGE="docker.io/grafana/grafana:10.2.1"
+GRAFANA_SIDECAR_FULL_IMAGE="quay.io/kiwigrid/k8s-sidecar:1.25.2"
+ADMWEBHOOK_FULL_IMAGE="registry.k8s.io/ingress-nginx/kube-webhook-certgen:v20221220-controller-v1.5.1-58-g787ea74b6"
+KSM_FULL_IMAGE="registry.k8s.io/kube-state-metrics/kube-state-metrics:v2.10.0"
+NODEXPORT_FULL_IMAGE="quay.io/prometheus/node-exporter:v1.7.0"
+PROMETHEUS_FULL_IMAGE="quay.io/prometheus/prometheus:v2.47.1"
+PROMOP_FULL_IMAGE="quay.io/prometheus-operator/prometheus-operator:v0.69.1"
+CONFIGRELOAD_FULL_IMAGE="quay.io/prometheus-operator/prometheus-config-reloader:v0.69.1"
+
+doitall "$PROMOP_FULL_IMAGE"          "TEMPFILE"  "monitoring/prom-operator_container_image.template"
+doitall "$ALERTMANAGER_FULL_IMAGE"    "TEMPFILE"  "ALERTMANAGER_"
+doitall "$ADMWEBHOOK_FULL_IMAGE"      "TEMPFILE"  "ADMWEBHOOK_"
+doitall "$KSM_FULL_IMAGE"             "TEMPFILE"  "KSM_"
+doitall "$NODEXPORT_FULL_IMAGE"       "TEMPFILE"  "NODEXPORT_"
+doitall "$PROMETHEUS_FULL_IMAGE"      "TEMPFILE"  "PROMETHEUS_"
+doitall "$CONFIGRELOAD_FULL_IMAGE"    "TEMPFILE"  "CONFIGRELOAD_"
+doitall "$GRAFANA_FULL_IMAGE"         "TEMPFILE"  "GRAFANA_"
+doitall "$GRAFANA_SIDECAR_FULL_IMAGE" "TEMPFILE"  "SIDECAR_"
+
 
 
 #### REMOVE FOLLOWING
